Update xxx to edu
