+++ sobit_edu_bringup/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>sobit_xxx_bringup</name>
+  <name>sobit_edu_bringup</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
   <maintainer email="sobits@todo.todo">sobits</maintainer>
